+++ packages/cli/src/commands/GenerateCacheCommand.ts
@@ -30,7 +30,7 @@ export class GenerateCacheCommand implements BaseCommand<CacheArgs> {
       metadataCache: { enabled: true, adapter: FileCacheAdapter, options },
     });
 
-    config.getMetadataCacheAdapter().clear();
+    await config.getMetadataCacheAdapter().clear();
     config.set('logger', CLIHelper.dump.bind(null));
     config.set('debug', true);
     const discovery = new MetadataDiscovery(MetadataStorage.init(), config.getDriver().getPlatform(), config);