+++ packages/knex/src/AbstractSqlDriver.ts
@@ -376,7 +376,8 @@ export abstract class AbstractSqlDriver<Connection extends AbstractSqlConnection
           }
         });
 
-      const targetProps = ref
+      const mapToPk = !!(ref || prop.mapToPk);
+      const targetProps = mapToPk
         ? meta2.getPrimaryProps()
         : meta2.props.filter(prop => this.platform.shouldHaveColumn(prop, hint.children as any || []));
       const tz = this.platform.getTimezone();
@@ -419,7 +420,7 @@ export abstract class AbstractSqlDriver<Connection extends AbstractSqlConnection
         prop.fieldNames.map(name => delete root![`${relationAlias}__${name}` as EntityKey<T>]);
       }
 
-      if (ref) {
+      if (mapToPk) {
         const tmp = Object.values(relationPojo);
         /* istanbul ignore next */
         relationPojo = (meta2.compositePK ? tmp : tmp[0]) as EntityData<T>;
@@ -1113,6 +1114,11 @@ export abstract class AbstractSqlDriver<Connection extends AbstractSqlConnection
         return true;
       }
 
+      // skip redundant joins for 1:1 owner population hints when using `mapToPk`
+      if (prop.kind === ReferenceKind.ONE_TO_ONE && prop.mapToPk && prop.owner) {
+        return false;
+      }
+
       if ((options?.strategy || hint.strategy || prop.strategy || this.config.get('loadStrategy')) !== LoadStrategy.JOINED) {
         // force joined strategy for explicit 1:1 owner populate hint as it would require a join anyway
         return prop.kind === ReferenceKind.ONE_TO_ONE && !prop.owner;
@@ -1243,9 +1249,9 @@ export abstract class AbstractSqlDriver<Connection extends AbstractSqlConnection
 
       const childExclude = exclude ? Utils.extractChildElements(exclude as string[], prop.name) : exclude;
 
-      if (!ref) {
+      if (!ref && !prop.mapToPk) {
         fields.push(...this.getFieldsForJoinedLoad(qb, meta2, childExplicitFields.length === 0 ? undefined : childExplicitFields, childExclude, hint.children as any, options, tableAlias, path, count));
-      } else if (hint.filter) {
+      } else if (hint.filter || prop.mapToPk) {
         fields.push(...prop.referencedColumnNames!.map(col => qb.helper.mapper(`${tableAlias}.${col}`, qb.type, undefined, `${tableAlias}__${col}`)));
       }
     }